+++ tests/test_data_migrations.py
@@ -19,7 +19,13 @@ from starlette.responses import StreamingResponse
 
 from cadwyn import VersionedAPIRouter, generate_code_for_versioned_packages
 from cadwyn._compat import PYDANTIC_V2, model_dump
-from cadwyn.exceptions import CadwynError, CadwynHeadRequestValidationError
+from cadwyn.exceptions import (
+    CadwynError,
+    CadwynHeadRequestValidationError,
+    RouteByPathConverterDoesNotApplyToAnythingError,
+    RouteRequestBySchemaConverterDoesNotApplyToAnythingError,
+    RouteResponseBySchemaConverterDoesNotApplyToAnythingError,
+)
 from cadwyn.route_generation import InternalRepresentationOf
 from cadwyn.structure import (
     VersionChange,
@@ -191,7 +197,7 @@ def version_change_2(head_module):
     )
 
 
-@pytest.fixture(params=["no request", "with request"])
+@pytest.fixture(params=["without_request", "with request"])
 def _post_endpoint_with_extra_depends(  # noqa: PT005
     request: pytest.FixtureRequest,
     router: VersionedAPIRouter,
@@ -199,7 +205,7 @@ def _post_endpoint_with_extra_depends(  # noqa: PT005
     head_module: ModuleType,
     _post_endpoint: Callable[..., Coroutine[Any, Any, dict[str, Any]]],  # pyright: ignore[reportRedeclaration]
 ):
-    if request.param == "no request":
+    if request.param == "without_request":
         router.routes = []
 
         @router.post(test_path)
@@ -319,26 +325,6 @@ class TestRequestMigrations:
         assert clients[date(2001, 1, 1)].get(test_path, headers={"my-header": "wow"}).json() == 83
         with pytest.raises(CadwynHeadRequestValidationError):
             clients[date(2000, 1, 1)].get(test_path, headers={"my-header": "wow"}).json()
-
-    def test__optional_body_field(
-        self,
-        create_versioned_clients: CreateVersionedClients,
-        head_module: ModuleType,
-        test_path: Literal["/test"],
-        router: VersionedAPIRouter,
-    ):
-        @router.post(test_path)
-        async def route(payload: head_module.AnyRequestSchema | None = Body(None)):
-            return payload or {"hello": "world"}
-
-        @convert_request_to_next_version_for(head_module.AnyRequestSchema)
-        def migrator(request: RequestInfo):
-            assert request.body is None
-
-        clients = create_versioned_clients(version_change(migrator=migrator))
-
-        assert clients[date(2000, 1, 1)].post(test_path).json() == {"hello": "world"}
-        assert clients[date(2001, 1, 1)].post(test_path).json() == {"hello": "world"}
 
     def test__internal_schema_specified__with_no_migrations__body_gets_parsed_to_internal_request_schema(
         self,
@@ -964,42 +950,6 @@ class TestHowAndWhenMigrationsApply:
             == []
         )
 
-    def test__migrate_one_version_down__with_inapplicable_migrations__result_is_only_affected_by_applicable_migrations(
-        self,
-        version_change_1: type[VersionChange],
-        create_versioned_clients: CreateVersionedClients,
-        test_path: Literal["/test"],
-        _post_endpoint,
-        head_module,
-    ):
-        def bad_req(request: RequestInfo):
-            raise NotImplementedError("I was not supposed to be ever called! This is very bad!")
-
-        def bad_resp(response: ResponseInfo):
-            raise NotImplementedError("I was not supposed to be ever called! This is very bad!")
-
-        clients = create_versioned_clients(
-            [
-                version_change_1,
-                version_change(
-                    wrong_body_schema=convert_request_to_next_version_for(head_module.AnyResponseSchema)(bad_req),
-                    wrong_resp_schema=convert_response_to_previous_version_for(head_module.AnyRequestSchema)(
-                        bad_resp,
-                    ),
-                    wrong_req_path=convert_request_to_next_version_for("/wrong_path", ["POST"])(bad_req),
-                    wrong_req_method=convert_request_to_next_version_for(test_path, ["GET"])(bad_req),
-                    wrong_resp_path=convert_response_to_previous_version_for("/wrong_path", ["POST"])(bad_resp),
-                    wrong_resp_method=convert_response_to_previous_version_for(test_path, ["GET"])(bad_resp),
-                ),
-            ],
-        )
-        assert len(clients) == 2
-        assert clients[date(2000, 1, 1)].post(test_path, json=[]).json()["body"] == [
-            "request change 1",
-            "response change 1",
-        ]
-        assert clients[date(2001, 1, 1)].post(test_path, json=[]).json()["body"] == []
-
     def test__cookies_can_be_deleted_during_migrations(
         self,
         create_versioned_clients: CreateVersionedClients,
@@ -1274,6 +1224,82 @@ def test__request_and_response_migrations__for_endpoint_with_modified_status_cod
     assert resp_2001.json() == 83
 
 
+@pytest.mark.parametrize(("path", "method"), [("/NOT_test", "POST"), ("/test", "PUT")])
+def test__request_by_path_migration__for_nonexistent_endpoint_path__should_raise_error(
+    create_versioned_clients: CreateVersionedClients,
+    head_module,
+    router: VersionedAPIRouter,
+    path: str,
+    method: str,
+):
+    @router.post("/test")
+    async def endpoint():
+        raise NotImplementedError
+
+    @convert_request_to_next_version_for(path, [method])
+    def request_converter(request: RequestInfo):
+        raise NotImplementedError
+
+    with pytest.raises(RouteByPathConverterDoesNotApplyToAnythingError):
+        create_versioned_clients(version_change(converter=request_converter))
+
+
+@pytest.mark.parametrize(("path", "method"), [("/NOT_test", "POST"), ("/test", "PUT")])
+def test__response_by_path_migration__for_nonexistent_endpoint_path__should_raise_error(
+    create_versioned_clients: CreateVersionedClients,
+    head_module,
+    router: VersionedAPIRouter,
+    path: str,
+    method: str,
+):
+    @router.post("/test")
+    async def endpoint():
+        raise NotImplementedError
+
+    @convert_response_to_previous_version_for(path, [method])
+    def response_converter(response: ResponseInfo):
+        raise NotImplementedError
+
+    with pytest.raises(RouteByPathConverterDoesNotApplyToAnythingError):
+        create_versioned_clients(version_change(converter=response_converter))
+
+
+def test__request_by_schema_migration__for_nonexistent_schema__should_raise_error(
+    create_versioned_clients: CreateVersionedClients,
+    head_module,
+    router: VersionedAPIRouter,
+):
+    @router.post("/test", response_model=head_module.AnyResponseSchema)
+    async def endpoint(body: head_module.AnyRequestSchema):
+        raise NotImplementedError
+
+    # Using response model for requests to cause an error
+    @convert_request_to_next_version_for(head_module.AnyResponseSchema)
+    def request_converter(request: RequestInfo):
+        raise NotImplementedError
+
+    with pytest.raises(RouteRequestBySchemaConverterDoesNotApplyToAnythingError):
+        create_versioned_clients(version_change(converter=request_converter))
+
+
+def test__response_by_schema_migration__for_nonexistent_schema__should_raise_error(
+    create_versioned_clients: CreateVersionedClients,
+    head_module,
+    router: VersionedAPIRouter,
+):
+    @router.post("/test", response_model=head_module.AnyResponseSchema)
+    async def endpoint(body: head_module.AnyRequestSchema):
+        raise NotImplementedError
+
+    # Using request model for responses to cause an error
+    @convert_response_to_previous_version_for(head_module.AnyRequestSchema)
+    def response_converter(response: ResponseInfo):
+        raise NotImplementedError
+
+    with pytest.raises(RouteResponseBySchemaConverterDoesNotApplyToAnythingError):
+        create_versioned_clients(version_change(converter=response_converter))
+
+
 def test__manual_response_migrations(
     head_with_empty_classes: _FakeModuleWithEmptyClasses,
     head_package_path: str,
@@ -1352,17 +1378,17 @@ def test__request_and_response_migrations__with_multiple_schemas_in_converters(
     )
 
     @router.post("/test_1")
-    async def endpoint_1(body: latest.Request_1) -> latest.Response_1:  # pyright: ignore[reportInvalidTypeForm]
+    async def endpoint_1(body: latest.Request_1) -> latest.Response_1:
         body.i.append("test_1")
         return body
 
     @router.post("/test_2")
-    async def endpoint_2(body: latest.Request_2) -> latest.Response_2:  # pyright: ignore[reportInvalidTypeForm]
+    async def endpoint_2(body: latest.Request_2) -> latest.Response_2:
         body.i.append("test_2")
         return body
 
     @router.post("/test_3")
-    async def endpoint_3(body: latest.Request_3) -> latest.Response_3:  # pyright: ignore[reportInvalidTypeForm]
+    async def endpoint_3(body: latest.Request_3) -> latest.Response_3:
         body.i.append("test_3")
         return body